[Paywalls V2] File organization restructure (#4497)
* Moved files around and split up main paywalls v2 view

* Reworked a view more things

* So much rework but cleaner

* This feels cleaner
