+++ RevenueCatUI/PaywallView.swift
@@ -254,7 +254,7 @@ public struct PaywallView: View {
 
         #if PAYWALL_COMPONENTS
         if let componentData = offering.paywallComponentsData {
-            TemplateComponentsView(
+            PaywallsV2View(
                 paywallComponentsData: componentData,
                 offering: offering,
                 introEligibilityChecker: .default(),
