+++ RevenueCatUI/Templates/V2/Components/Image/ImageComponentViewModel.swift
