+++ RevenueCatUI/Templates/V2/Components/Root/RootViewModel.swift
