+++ RevenueCatUI/Templates/V2/Components/StickyFooter/StickyFooterComponentViewModel.swift
